step 3: Implement dynamic proof reward (#199)
This PR implement a dynamic proof reward. When the block's proving time `d` is bigger than 1.25x the moving average `a`, the block fee will start to increase and when `d == 4*a`, the block fee drops to the minimal m = `4*f` where f is the latest base fee.
![WechatIMG306](https://user-images.githubusercontent.com/99078276/198292471-8d3424b3-1237-4fcb-87a0-e228e56fd482.jpeg)


This means the proof reward can only become larger when new proof time get larger, which indicates that provers are reluctant to prove a block most likely because `f` is too small.

When the proof time gets smaller, we do not decrease the base fee, which allows encourage provers to submit proofs as soon as possible to finalize blocks earlier.