The section on What does Taiko mean
The description wass not clear. E.g. how does  一鼓作气 (Yīgǔzuòqì) connect to drum (Taiko)? I have tried to change the description to make it clearer. 

I wanted to add one more sentence or observation on the Chinese character and the Japanese character for a drum being the same (鼓). However, it is not relevant for our curent requirements so I have not added it. Let me know if this change makes sense.